refactor: extract candidate selector into own component
