+++ src/pages/ExamRoot.tsx
@@ -2,7 +2,7 @@ import { Section } from '@radix-ui/themes'
 import React, { FC } from 'react'
 import { Outlet } from 'react-router-dom'
 
-import ExamNavBar from '../components/ExamNavBar'
+import ExamNavBar from '../components/topBars/ExamNavBar'
 
 const ExamRoot: FC = () => {
   return (
